Refactor database client implementation and update import paths
