+++ frontend/app/api/create-team/route.ts
@@ -7,6 +7,7 @@ import { base } from 'viem/chains';
 import { chain } from '@/config/chains';
 import { BaseError, ContractFunctionRevertedError } from 'viem';
 import { sendWebhookMessage } from '@/lib/webhook';
+import { checkAuthSignatureAndMessage } from '@/lib/auth';
 
 
 interface CreateTeamRequest {
@@ -24,11 +25,21 @@ interface CreateTeamRequest {
 export async function POST(req: NextRequest): Promise<NextResponse> {
     try {
         const body: CreateTeamRequest = await req.json();
-        const { teamName, countryId, walletAddress } = body;
+        const { teamName, countryId, walletAddress, signature, message } = body;
 
-        // Authentication is handled by Next.js middleware
-        // walletAddress is already validated by middleware
-        // Sentry user context is set in middleware
+        // Verify signature (middleware validates format, we verify here in Node.js runtime)
+        const { isValid, error: authError } = await checkAuthSignatureAndMessage(
+            signature,
+            message,
+            walletAddress
+        );
+
+        if (!isValid) {
+            return NextResponse.json(
+                { success: false, error: authError || 'Signature verification failed' },
+                { status: 401 }
+            );
+        }
 
         // Log the received data to console
         console.log('=== API RECEIVED AUTHENTICATION DATA ===');
