+++ frontend/app/game/[gameId]/page.tsx
@@ -37,8 +37,25 @@ import GameResultModal from './components/GameResultModal';
 import CancelGameModal from './components/CancelGameModal';
 
 export default function GamePage() {
-    const params = useParams();
-    const gameId = params.gameId as string;
+    const params = useParams<{ gameId?: string }>();
+    const gameId = params?.gameId ?? '';
+
+    if (!gameId) {
+        return (
+            <div className="min-h-screen flex items-center justify-center bg-green-100 p-4">
+                <div className="max-w-md text-center bg-white shadow-sm rounded-lg p-6 border border-gray-100">
+                    <h1 className="text-xl font-semibold text-gray-800 mb-2">Game not found</h1>
+                    <p className="text-gray-600 mb-4">Please return to the lobby and select a valid game.</p>
+                    <a
+                        href="/"
+                        className="inline-flex items-center justify-center px-4 py-2 bg-green-600 text-white rounded hover:bg-green-700 transition-colors"
+                    >
+                        Back to Home
+                    </a>
+                </div>
+            </div>
+        );
+    }
 
     // Wallet connection
     const { address, isConnected } = useAccount();
